feat: add auth on vote api

+++ server/src/votes/votes.service.ts
@@ -1,14 +1,21 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { CreateVoteDto } from './dto/create-vote.dto';
 import { UpdateVoteDto } from './dto/update-vote.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { UserStorage } from 'src/users/user.store';
 
 @Injectable()
 export class VotesService {
   constructor(private prisma: PrismaService) {}
 
   create(createVoteDto: CreateVoteDto) {
-    return this.prisma.vote.create({ data: createVoteDto });
+    const user = UserStorage.get();
+    return this.prisma.vote.create({
+      data: {
+        ...createVoteDto,
+        authorId: user.id,
+      },
+    });
   }
 
   findAll() {
@@ -24,14 +31,25 @@ export class VotesService {
     });
   }
 
-  update(id: number, updateVoteDto: UpdateVoteDto) {
+  async checkVoteAuthor(voteId) {
+    const user = UserStorage.get();
+    const vote = await this.prisma.vote.findUnique({ where: { id: voteId } });
+    if (!vote.authorId || vote.authorId !== user.id) {
+      throw new UnauthorizedException('user is not the vote author');
+    }
+    return;
+  }
+
+  async update(id: number, updateVoteDto: UpdateVoteDto) {
+    await this.checkVoteAuthor(id);
     return this.prisma.vote.update({
       where: { id },
       data: updateVoteDto,
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    await this.checkVoteAuthor(id);
     return this.prisma.vote.delete({ where: { id } });
   }
 }
